+++ FECore/FESolidDomain.cpp
@@ -46,7 +46,7 @@ bool FESolidDomain::Init()
 
 	// init solid element data
 	double Ji[3][3];
-	for (size_t i = 0; i<m_Elem.size(); ++i)
+	for (int i = 0; i<(int)m_Elem.size(); ++i)
 	{
 		FESolidElement& el = Element(i);
 		int nint = el.GaussPoints();
@@ -65,7 +65,7 @@ bool FESolidDomain::Init()
 // Reset data
 void FESolidDomain::Reset()
 {
-    for (size_t i=0; i<m_Elem.size(); ++i)
+	for (int i = 0; i<(int)m_Elem.size(); ++i)
     {
         FESolidElement& el = Element(i);
         int nint = el.GaussPoints();
